+++ ship/TIELN.cpp
@@ -1,7 +1,7 @@
 #include "TIELN.hpp"
 
 const std::string TIELN::MODEL = "TIE/LN";
-const unsigned long long TIELN::WEIGHT = 6;
+const double TIELN::WEIGHT = 6;
 const unsigned TIELN::SPEED = 100;
 unsigned TIELN::id = 0;
